pyproject.toml
[project]
name = "ai-de-pipeline-proj-1"
version = "0.1.0"
description = "Data pipeline for USGS Earthquake catalog data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dlt[bigquery,duckdb,filesystem,gs]>=1.9.0",
    "requests>=2.31.0",
    "pyarrow>=15.0.0",
    "python-dotenv>=1.0.0",
    "google-cloud-bigquery-storage>=2.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/earthquake_pipeline"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
]

[dependency-groups]
dev = [
    "enlighten>=1.14.1",
    "pre-commit>=4.2.0",
    "ruff>=0.11.2",
    "streamlit>=1.44.0",
]
